+++ lib/zendesk2/categories.rb
