+++ lib/zendesk2/client.rb
@@ -8,21 +8,21 @@ require 'zendesk2/help_center'
 require 'zendesk2/search'
 require 'zendesk2/audit_event'
 require 'zendesk2/category'
-require 'zendesk2/collections/categories'
+require 'zendesk2/categories'
 require 'zendesk2/create_category'
 require 'zendesk2/destroy_category'
 require 'zendesk2/get_categories'
 require 'zendesk2/get_category'
 require 'zendesk2/update_category'
 require 'zendesk2/forum'
-require 'zendesk2/collections/forums'
+require 'zendesk2/forums'
 require 'zendesk2/create_forum'
 require 'zendesk2/destroy_forum'
 require 'zendesk2/get_forums'
 require 'zendesk2/get_forum'
 require 'zendesk2/update_forum'
 require 'zendesk2/group'
-require 'zendesk2/collections/groups'
+require 'zendesk2/groups'
 require 'zendesk2/create_group'
 require 'zendesk2/destroy_group'
 require 'zendesk2/get_groups'
@@ -30,7 +30,7 @@ require 'zendesk2/get_group'
 require 'zendesk2/update_group'
 require 'zendesk2/get_assignable_groups'
 require 'zendesk2/user'
-require 'zendesk2/collections/users'
+require 'zendesk2/users'
 require 'zendesk2/create_user'
 require 'zendesk2/destroy_user'
 require 'zendesk2/get_users'
@@ -42,7 +42,7 @@ require 'zendesk2/mark_user_identity_primary'
 require 'zendesk2/get_user_memberships'
 require 'zendesk2/get_user_organizations'
 require 'zendesk2/ticket'
-require 'zendesk2/collections/tickets'
+require 'zendesk2/tickets'
 require 'zendesk2/create_ticket'
 require 'zendesk2/destroy_ticket'
 require 'zendesk2/get_tickets'
@@ -51,34 +51,34 @@ require 'zendesk2/update_ticket'
 require 'zendesk2/get_requested_tickets'
 require 'zendesk2/get_ccd_tickets'
 require 'zendesk2/ticket_audit'
-require 'zendesk2/collections/ticket_audits'
+require 'zendesk2/ticket_audits'
 require 'zendesk2/get_ticket_audits'
 require 'zendesk2/get_ticket_audit'
 require 'zendesk2/ticket_field'
-require 'zendesk2/collections/ticket_fields'
+require 'zendesk2/ticket_fields'
 require 'zendesk2/create_ticket_field'
 require 'zendesk2/destroy_ticket_field'
 require 'zendesk2/get_ticket_fields'
 require 'zendesk2/get_ticket_field'
 require 'zendesk2/update_ticket_field'
 require 'zendesk2/topic'
-require 'zendesk2/collections/topics'
+require 'zendesk2/topics'
 require 'zendesk2/create_topic'
 require 'zendesk2/destroy_topic'
 require 'zendesk2/get_topics'
 require 'zendesk2/get_topic'
 require 'zendesk2/update_topic'
 require 'zendesk2/topic_comment'
-require 'zendesk2/collections/topic_comments'
+require 'zendesk2/topic_comments'
 require 'zendesk2/create_topic_comment'
 require 'zendesk2/destroy_topic_comment'
 require 'zendesk2/get_topic_comments'
 require 'zendesk2/get_topic_comment'
 require 'zendesk2/ticket_comment'
-require 'zendesk2/collections/ticket_comments'
+require 'zendesk2/ticket_comments'
 require 'zendesk2/get_ticket_comments'
 require 'zendesk2/organization'
-require 'zendesk2/collections/organizations'
+require 'zendesk2/organizations'
 require 'zendesk2/create_organization'
 require 'zendesk2/destroy_organization'
 require 'zendesk2/get_organizations'
@@ -90,28 +90,28 @@ require 'zendesk2/get_organization_by_external_id'
 require 'zendesk2/get_organization_memberships'
 require 'zendesk2/search_organization'
 require 'zendesk2/user_field'
-require 'zendesk2/collections/user_fields'
+require 'zendesk2/user_fields'
 require 'zendesk2/create_user_field'
 require 'zendesk2/destroy_user_field'
 require 'zendesk2/get_user_fields'
 require 'zendesk2/get_user_field'
 require 'zendesk2/update_user_field'
 require 'zendesk2/user_identity'
-require 'zendesk2/collections/user_identities'
+require 'zendesk2/user_identities'
 require 'zendesk2/create_user_identity'
 require 'zendesk2/destroy_user_identity'
 require 'zendesk2/get_user_identities'
 require 'zendesk2/get_user_identity'
 require 'zendesk2/update_user_identity'
 require 'zendesk2/membership'
-require 'zendesk2/collections/memberships'
+require 'zendesk2/memberships'
 require 'zendesk2/create_membership'
 require 'zendesk2/destroy_membership'
 require 'zendesk2/get_memberships'
 require 'zendesk2/get_membership'
 require 'zendesk2/mark_membership_default'
 require 'zendesk2/view'
-require 'zendesk2/collections/views'
+require 'zendesk2/views'
 require 'zendesk2/create_view'
 require 'zendesk2/destroy_view'
 require 'zendesk2/get_views'
