+++ lib/zendesk2/ticket_metrics.rb
