+++ lib/zendesk2/topic_comments.rb
