+++ lib/zendesk2/user_identities.rb
