+++ lib/zendesk2/users.rb
